requirements.txt
africastalking==1.2.9
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
josepy==2.1.0
mozilla-django-oidc==4.0.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
schema==0.7.7
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
